# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/curd/filetemplate/add": {
            "post": {
                "description": "新增",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件模版管理"
                ],
                "summary": "新增",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Add FileTemplate",
                        "name": "account",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/curd.FileTemplateInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/curd.CommonResponse"
                        }
                    }
                }
            }
        },
        "/api/curd/filetemplate/all": {
            "get": {
                "description": "查询所有",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件模版管理"
                ],
                "summary": "查询所有",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/curd.GetAllFileTemplateResponse"
                        }
                    }
                }
            }
        },
        "/api/curd/filetemplate/copy": {
            "post": {
                "description": "复制",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件模版管理"
                ],
                "summary": "复制",
                "parameters": [
                    {
                        "description": "Copy FileTemplate",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/curd.DelRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/curd.CommonResponse"
                        }
                    }
                }
            }
        },
        "/api/curd/filetemplate/delete": {
            "delete": {
                "description": "删除",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件模版管理"
                ],
                "summary": "删除",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Delete FileTemplate",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/curd.DelRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/curd.CommonResponse"
                        }
                    }
                }
            }
        },
        "/api/curd/filetemplate/detail": {
            "get": {
                "description": "查询明细",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件模版管理"
                ],
                "summary": "查询明细",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/curd.GetFileTemplateDetailResponse"
                        }
                    }
                }
            }
        },
        "/api/curd/filetemplate/export": {
            "get": {
                "description": "导出",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "文件模版管理"
                ],
                "summary": "导出",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "从1开始",
                        "name": "pageIndex",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "默认每页10条",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序字段",
                        "name": "orderField",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "是否倒序排序",
                        "name": "desc",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "文件名称",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "description": "IDs",
                        "name": "ids",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/curd.CommonResponse"
                        }
                    }
                }
            }
        },
        "/api/curd/filetemplate/import": {
            "post": {
                "description": "导入",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件模版管理"
                ],
                "summary": "导入",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer+空格+Token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "要上传的文件",
                        "name": "files",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/curd.CommonResponse"
                        }
                    }
                }
            }
        },
        "/api/curd/filetemplate/query": {
            "get": {
                "description": "分页查询",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "文件模版管理"
                ],
                "summary": "分页查询",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "从1开始",
                        "name": "pageIndex",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "默认每页10条",
                        "name": "pageSize",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序字段",
                        "name": "orderField",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "是否倒序排序",
                        "name": "desc",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "文件名称",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/curd.QueryFileTemplateResponse"
                        }
                    }
                }
            }
        },
        "/api/curd/filetemplate/update": {
            "put": {
                "description": "更新",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件模版管理"
                ],
                "summary": "更新",
                "parameters": [
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Update FileTemplate",
                        "name": "account",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/curd.FileTemplateInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/curd.CommonResponse"
                        }
                    }
                }
            }
        },
...
        "/api/curd/template/copy": {
            "post": {
                "description": "复制",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "代码模板管理"
                ],
                "summary": "复制",
                "parameters": [
                    {
                        "description": "Copy Template",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/curd.DelRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "jwt token",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/curd.CommonResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "curd.CommonResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "curd.DelRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "tenantID": {
                    "type": "string"
                }
            }
        },
...
        "curd.FileTemplateInfo": {
            "type": "object",
            "properties": {
                "body": {
                    "description": "一组模板ID",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "dir": {
                    "description": "文件存放的目录",
                    "type": "string"
                },
                "end": {
                    "description": "结尾模板ID",
                    "type": "string"
                },
                "fileNameSuffix": {
                    "description": "文件名称后缀",
                    "type": "string"
                },
                "group": {
                    "description": "分组",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "language": {
                    "description": "编程语言",
                    "type": "string"
                },
                "name": {
                    "description": "文件模版名称",
                    "type": "string"
                },
                "package": {
                    "description": "包名",
                    "type": "string"
                },
                "params": {
                    "description": "自定义参数",
                    "type": "string"
                },
                "start": {
                    "description": "开头模板ID",
                    "type": "string"
                }
            }
        },
...
        "curd.GetAllFileTemplateResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/curd.FileTemplateInfo"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "curd.GetFileTemplateDetailResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/curd.FileTemplateInfo"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "curd.QueryFileTemplateResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/curd.FileTemplateInfo"
                    }
                },
                "message": {
                    "type": "string"
                },
                "pages": {
                    "type": "integer"
                },
                "records": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
    }
}